// srcold/interfaces/xiu.xian.ts
export class XiuXian {
	currentMajorLevel: number = 0;

	currentMinorLevel: number = 0;

	nextMajorLevel: number = 0;

	nextMinorLevel: number = 0;

	currentMajorLevelText: string = "";

	currentMinorLevelText: string = "";

	nextMajorLevelText: string = "";

	nextMinorLevelText: string = "";

	totalLeft: number = 0;

	leftSecond: number = 0;

	leftMinute: number = 0;

	leftHour: number = 0;

	isMaxLevel: boolean = false;

	isNextMaxLevel: boolean = false;

	totalExperience: number = 0;

	levelColor: string = "";

	expire: string;
}